refactor: split uiBuilderStore.ts into types, presets, and store modules
Extract 28 type/interface definitions into uiBuilderTypes.ts and
preset widget creation (7 screen presets) into uiBuilderPresets.ts.
Store re-exports all types for backward compatibility. Breaks circular
dependency by pointing widgetDefaults.ts at uiBuilderTypes directly.

1522 lines -> 1004 (store) + 219 (types) + 321 (presets)
