+++ web/src/stores/uiBuilderStore.ts
@@ -1,224 +1,47 @@
 import { create } from 'zustand';
+import { createPresetWidgets } from './uiBuilderPresets';
+import { getWidgetDefaults, DEFAULT_WIDGET_STYLE } from '@/components/editor/ui-builder/widgetDefaults';
 
-// ========== Type Definitions ==========
+// Re-export all types so existing consumers don't break
+export type {
+  WidgetType,
+  WidgetAnchor,
+  WidgetStyle,
+  DataBinding,
+  BindingTransform,
+  TextWidgetConfig,
+  ImageWidgetConfig,
+  ButtonAction,
+  ButtonWidgetConfig,
+  ProgressBarWidgetConfig,
+  PanelWidgetConfig,
+  GridWidgetConfig,
+  ScrollViewWidgetConfig,
+  SliderWidgetConfig,
+  ToggleWidgetConfig,
+  MinimapWidgetConfig,
+  UIWidgetBase,
+  UIWidget,
+  ScreenTransition,
+  UIScreen,
+  UITheme,
+  GameUIData,
+  ScreenPreset,
+  UIUndoAction,
+} from './uiBuilderTypes';
 
-export type WidgetType =
-  | 'text'
-  | 'image'
-  | 'button'
-  | 'progress_bar'
-  | 'panel'
-  | 'grid'
-  | 'scroll_view'
-  | 'slider'
-  | 'toggle'
-  | 'minimap';
-
-export type WidgetAnchor =
-  | 'top_left' | 'top_center' | 'top_right'
-  | 'center_left' | 'center' | 'center_right'
-  | 'bottom_left' | 'bottom_center' | 'bottom_right';
-
-export interface WidgetStyle {
-  backgroundColor: string | null;
-  borderWidth: number;
-  borderColor: string;
-  borderRadius: number;
-  padding: [number, number, number, number];
-  opacity: number;
-  overflow: 'visible' | 'hidden' | 'scroll';
-  fontFamily: string;
-  fontSize: number;
-  fontWeight: 'normal' | 'bold';
-  color: string;
-  textAlign: 'left' | 'center' | 'right';
-  lineHeight: number;
-  textShadow: string | null;
-  rotation: number;
-  scaleX: number;
-  scaleY: number;
-}
-
-export interface DataBinding {
-  stateKey: string;
-  direction: 'read' | 'write' | 'read_write';
-  transform: BindingTransform | null;
-}
-
-export type BindingTransform =
-  | { type: 'format'; template: string }
-  | { type: 'map'; entries: Array<{ from: unknown; to: string }> }
-  | { type: 'clamp'; min: number; max: number }
-  | { type: 'multiply'; factor: number }
-  | { type: 'round'; decimals: number };
-
-export interface TextWidgetConfig {
-  content: string;
-  binding: DataBinding | null;
-}
-
-export interface ImageWidgetConfig {
-  assetId: string | null;
-  src: string | null;
-  fit: 'contain' | 'cover' | 'fill' | 'none';
-  alt: string;
-}
-
-export type ButtonAction =
-  | { type: 'show_screen'; screenId: string }
-  | { type: 'hide_screen'; screenId: string }
-  | { type: 'toggle_screen'; screenId: string }
-  | { type: 'set_state'; key: string; value: unknown }
-  | { type: 'call_function'; functionName: string }
-  | { type: 'scene_reset' }
-  | { type: 'none' };
-
-export interface ButtonWidgetConfig {
-  label: string;
-  hoverStyle: Partial<WidgetStyle>;
-  activeStyle: Partial<WidgetStyle>;
-  action: ButtonAction;
-}
-
-export interface ProgressBarWidgetConfig {
-  valueBinding: DataBinding;
-  min: number;
-  max: number;
-  direction: 'horizontal' | 'vertical' | 'radial';
-  fillColor: string;
-  trackColor: string;
-  showLabel: boolean;
-  labelFormat: string;
-}
-
-export interface PanelWidgetConfig {
-  layout: 'free' | 'vertical' | 'horizontal';
-  gap: number;
-  alignItems: 'start' | 'center' | 'end' | 'stretch';
-}
-
-export interface GridWidgetConfig {
-  columns: number;
-  rows: number;
-  cellWidth: number;
-  cellHeight: number;
-  gap: number;
-  itemBinding: DataBinding | null;
-  cellTemplate: UIWidget | null;
-}
-
-export interface ScrollViewWidgetConfig {
-  direction: 'vertical' | 'horizontal' | 'both';
-  showScrollbar: boolean;
-  scrollbarColor: string;
-}
-
-export interface SliderWidgetConfig {
-  valueBinding: DataBinding;
-  min: number;
-  max: number;
-  step: number;
-  orientation: 'horizontal' | 'vertical';
-  trackColor: string;
-  thumbColor: string;
-  fillColor: string;
-}
-
-export interface ToggleWidgetConfig {
-  valueBinding: DataBinding;
-  onLabel: string;
-  offLabel: string;
-  trackColorOn: string;
-  trackColorOff: string;
-  thumbColor: string;
-}
-
-export interface MinimapWidgetConfig {
-  zoom: number;
-  trackedEntityIds: string[];
-  trackedEntityTags: string[];
-  dotColor: string;
-  dotSize: number;
-  playerDotColor: string;
-  backgroundColor: string;
-  showBorder: boolean;
-}
-
-export interface UIWidgetBase {
-  id: string;
-  name: string;
-  type: WidgetType;
-  x: number;
-  y: number;
-  width: number;
-  height: number;
-  anchor: WidgetAnchor;
-  style: WidgetStyle;
-  visible: boolean;
-  interactable: boolean;
-  parentWidgetId: string | null;
-  children: string[];
-}
-
-export type UIWidget = UIWidgetBase & (
-  | { type: 'text'; config: TextWidgetConfig }
-  | { type: 'image'; config: ImageWidgetConfig }
-  | { type: 'button'; config: ButtonWidgetConfig }
-  | { type: 'progress_bar'; config: ProgressBarWidgetConfig }
-  | { type: 'panel'; config: PanelWidgetConfig }
-  | { type: 'grid'; config: GridWidgetConfig }
-  | { type: 'scroll_view'; config: ScrollViewWidgetConfig }
-  | { type: 'slider'; config: SliderWidgetConfig }
-  | { type: 'toggle'; config: ToggleWidgetConfig }
-  | { type: 'minimap'; config: MinimapWidgetConfig }
-);
-
-export interface ScreenTransition {
-  type: 'none' | 'fade' | 'slide_left' | 'slide_right' | 'slide_up' | 'slide_down' | 'scale';
-  durationMs: number;
-  easing: 'linear' | 'ease_in' | 'ease_out' | 'ease_in_out';
-}
-
-export interface UIScreen {
-  id: string;
-  name: string;
-  widgets: UIWidget[];
-  visible: boolean;
-  showOnStart: boolean;
-  showOnKey: string | null;
-  transition: ScreenTransition;
-  zIndex: number;
-  backgroundColor: string;
-  blockInput: boolean;
-}
-
-export interface UITheme {
-  primaryColor: string;
-  secondaryColor: string;
-  backgroundColor: string;
-  textColor: string;
-  fontFamily: string;
-  fontSize: number;
-  borderRadius: number;
-}
-
-export interface GameUIData {
-  version: 1;
-  screens: UIScreen[];
-  globalTheme: UITheme | null;
-}
-
-export type ScreenPreset = 'blank' | 'hud' | 'main_menu' | 'pause_menu' | 'game_over' | 'inventory' | 'dialog';
-
-export type UIUndoAction =
-  | { type: 'add_screen'; screen: UIScreen }
-  | { type: 'delete_screen'; screen: UIScreen; index: number }
-  | { type: 'update_screen'; screenId: string; before: Record<string, unknown>; after: Record<string, unknown> }
-  | { type: 'add_widget'; screenId: string; widget: UIWidget }
-  | { type: 'delete_widget'; screenId: string; widget: UIWidget; index: number }
-  | { type: 'update_widget'; screenId: string; widgetId: string; before: Record<string, unknown>; after: Record<string, unknown> }
-  | { type: 'move_widget'; screenId: string; widgetId: string; beforeX: number; beforeY: number; afterX: number; afterY: number }
-  | { type: 'resize_widget'; screenId: string; widgetId: string; beforeW: number; beforeH: number; afterW: number; afterH: number };
+import type {
+  WidgetType,
+  WidgetStyle,
+  DataBinding,
+  UIWidgetBase,
+  UIWidget,
+  UIScreen,
+  UITheme,
+  GameUIData,
+  ScreenPreset,
+  UIUndoAction,
+} from './uiBuilderTypes';
 
 // ========== Store ==========
 
@@ -290,8 +113,6 @@ let idCounter = 0;
 function nextId(prefix: string): string {
   return `${prefix}_${Date.now()}_${++idCounter}`;
 }
-
-import { getWidgetDefaults, DEFAULT_WIDGET_STYLE } from '@/components/editor/ui-builder/widgetDefaults';
 
 export const useUIBuilderStore = create<UIBuilderState>((set, get) => ({
   screens: [],
@@ -1181,342 +1002,3 @@ export const useUIBuilderStore = create<UIBuilderState>((set, get) => ({
     }
   },
 }));
-
-// ========== Preset Helpers ==========
-
-function createPresetWidgets(preset: ScreenPreset): UIWidget[] {
-  const widgets: UIWidget[] = [];
-
-  switch (preset) {
-    case 'hud':
-      // Top-left score
-      widgets.push(createWidget('text', {
-        name: 'score_text',
-        x: 2,
-        y: 2,
-        width: 15,
-        height: 5,
-        config: { content: 'Score: {{score}}', binding: null },
-      }));
-      // Top-right health bar
-      widgets.push(createWidget('progress_bar', {
-        name: 'health_bar',
-        x: 80,
-        y: 2,
-        width: 18,
-        height: 3,
-        config: {
-          valueBinding: { stateKey: 'health', direction: 'read', transform: null },
-          min: 0,
-          max: 100,
-          direction: 'horizontal',
-          fillColor: '#22c55e',
-          trackColor: '#374151',
-          showLabel: true,
-          labelFormat: '{percent}%',
-        },
-      }));
-      // Bottom-center ammo text
-      widgets.push(createWidget('text', {
-        name: 'ammo_text',
-        x: 45,
-        y: 90,
-        width: 10,
-        height: 5,
-        config: { content: 'Ammo: {{ammo}}', binding: null },
-      }));
-      break;
-
-    case 'main_menu':
-      // Centered title
-      widgets.push(createWidget('text', {
-        name: 'title',
-        x: 35,
-        y: 20,
-        width: 30,
-        height: 10,
-        config: { content: 'Game Title', binding: null },
-        style: { fontSize: 48, textAlign: 'center' as const, fontWeight: 'bold' as const },
-      }));
-      // Play button
-      widgets.push(createWidget('button', {
-        name: 'play_button',
-        x: 40,
-        y: 45,
-        width: 20,
-        height: 6,
-        config: {
-          label: 'Play',
-          hoverStyle: { opacity: 0.8 },
-          activeStyle: { opacity: 0.6 },
-          action: { type: 'none' },
-        },
-      }));
-      // Settings button
-      widgets.push(createWidget('button', {
-        name: 'settings_button',
-        x: 40,
-        y: 55,
-        width: 20,
-        height: 6,
-        config: {
-          label: 'Settings',
-          hoverStyle: { opacity: 0.8 },
-          activeStyle: { opacity: 0.6 },
-          action: { type: 'none' },
-        },
-      }));
-      // Quit button
-      widgets.push(createWidget('button', {
-        name: 'quit_button',
-        x: 40,
-        y: 65,
-        width: 20,
-        height: 6,
-        config: {
-          label: 'Quit',
-          hoverStyle: { opacity: 0.8 },
-          activeStyle: { opacity: 0.6 },
-          action: { type: 'scene_reset' },
-        },
-      }));
-      break;
-
-    case 'pause_menu':
-      // Semi-transparent backdrop
-      widgets.push(createWidget('panel', {
-        name: 'backdrop',
-        x: 0,
-        y: 0,
-        width: 100,
-        height: 100,
-        config: { layout: 'free', gap: 0, alignItems: 'center' as const },
-        style: { backgroundColor: 'rgba(0,0,0,0.7)' },
-      }));
-      // Paused title
-      widgets.push(createWidget('text', {
-        name: 'paused_title',
-        x: 40,
-        y: 30,
-        width: 20,
-        height: 8,
-        config: { content: 'Paused', binding: null },
-        style: { fontSize: 36, textAlign: 'center' as const, fontWeight: 'bold' as const },
-      }));
-      // Resume button
-      widgets.push(createWidget('button', {
-        name: 'resume_button',
-        x: 40,
-        y: 45,
-        width: 20,
-        height: 6,
-        config: {
-          label: 'Resume',
-          hoverStyle: { opacity: 0.8 },
-          activeStyle: { opacity: 0.6 },
-          action: { type: 'none' },
-        },
-      }));
-      // Quit button
-      widgets.push(createWidget('button', {
-        name: 'quit_button',
-        x: 40,
-        y: 55,
-        width: 20,
-        height: 6,
-        config: {
-          label: 'Quit',
-          hoverStyle: { opacity: 0.8 },
-          activeStyle: { opacity: 0.6 },
-          action: { type: 'scene_reset' },
-        },
-      }));
-      break;
-
-    case 'game_over':
-      // Dark overlay
-      widgets.push(createWidget('panel', {
-        name: 'overlay',
-        x: 0,
-        y: 0,
-        width: 100,
-        height: 100,
-        config: { layout: 'free', gap: 0, alignItems: 'center' as const },
-        style: { backgroundColor: 'rgba(0,0,0,0.8)' },
-      }));
-      // Game Over title
-      widgets.push(createWidget('text', {
-        name: 'game_over_title',
-        x: 30,
-        y: 25,
-        width: 40,
-        height: 10,
-        config: { content: 'Game Over', binding: null },
-        style: { fontSize: 48, textAlign: 'center' as const, fontWeight: 'bold' as const, color: '#ef4444' },
-      }));
-      // Final score
-      widgets.push(createWidget('text', {
-        name: 'final_score',
-        x: 35,
-        y: 40,
-        width: 30,
-        height: 6,
-        config: { content: 'Score: {{score}}', binding: null },
-        style: { fontSize: 24, textAlign: 'center' as const },
-      }));
-      // Retry button
-      widgets.push(createWidget('button', {
-        name: 'retry_button',
-        x: 40,
-        y: 55,
-        width: 20,
-        height: 6,
-        config: {
-          label: 'Retry',
-          hoverStyle: { opacity: 0.8 },
-          activeStyle: { opacity: 0.6 },
-          action: { type: 'scene_reset' },
-        },
-      }));
-      // Main Menu button
-      widgets.push(createWidget('button', {
-        name: 'menu_button',
-        x: 40,
-        y: 65,
-        width: 20,
-        height: 6,
-        config: {
-          label: 'Main Menu',
-          hoverStyle: { opacity: 0.8 },
-          activeStyle: { opacity: 0.6 },
-          action: { type: 'none' },
-        },
-      }));
-      break;
-
-    case 'inventory':
-      // Inventory panel background
-      widgets.push(createWidget('panel', {
-        name: 'inventory_bg',
-        x: 10,
-        y: 5,
-        width: 80,
-        height: 90,
-        config: { layout: 'vertical', gap: 4, alignItems: 'stretch' as const },
-        style: { backgroundColor: 'rgba(24,24,27,0.95)', borderWidth: 1, borderColor: '#3f3f46' },
-      }));
-      // Title
-      widgets.push(createWidget('text', {
-        name: 'inventory_title',
-        x: 12,
-        y: 8,
-        width: 76,
-        height: 5,
-        config: { content: 'Inventory', binding: null },
-        style: { fontSize: 24, fontWeight: 'bold' as const },
-      }));
-      // Item grid
-      widgets.push(createWidget('grid', {
-        name: 'item_grid',
-        x: 12,
-        y: 16,
-        width: 76,
-        height: 65,
-        config: {
-          columns: 6,
-          rows: 4,
-          cellWidth: 64,
-          cellHeight: 64,
-          gap: 4,
-          itemBinding: null,
-          cellTemplate: null,
-        },
-      }));
-      // Close button
-      widgets.push(createWidget('button', {
-        name: 'close_button',
-        x: 82,
-        y: 6,
-        width: 6,
-        height: 4,
-        config: {
-          label: 'X',
-          hoverStyle: { opacity: 0.8 },
-          activeStyle: { opacity: 0.6 },
-          action: { type: 'none' },
-        },
-      }));
-      break;
-
-    case 'dialog':
-      // Dialog backdrop (bottom-center)
-      widgets.push(createWidget('panel', {
-        name: 'dialog_box',
-        x: 10,
-        y: 70,
-        width: 80,
-        height: 25,
-        config: { layout: 'vertical', gap: 2, alignItems: 'start' as const },
-        style: { backgroundColor: 'rgba(24,24,27,0.95)', borderWidth: 1, borderColor: '#3f3f46', borderRadius: 8 },
-      }));
-      // Speaker name
-      widgets.push(createWidget('text', {
-        name: 'speaker_name',
-        x: 12,
-        y: 72,
-        width: 20,
-        height: 4,
-        config: { content: 'Speaker', binding: null },
-        style: { fontSize: 16, fontWeight: 'bold' as const, color: '#60a5fa' },
-      }));
-      // Dialog text
-      widgets.push(createWidget('text', {
-        name: 'dialog_text',
-        x: 12,
-        y: 77,
-        width: 76,
-        height: 12,
-        config: { content: 'Dialog text goes here...', binding: null },
-        style: { fontSize: 14, lineHeight: 1.5 },
-      }));
-      // Continue indicator
-      widgets.push(createWidget('text', {
-        name: 'continue_hint',
-        x: 75,
-        y: 90,
-        width: 14,
-        height: 3,
-        config: { content: 'Click to continue ▶', binding: null },
-        style: { fontSize: 12, textAlign: 'right' as const, color: '#a1a1aa' },
-      }));
-      break;
-  }
-
-  return widgets;
-}
-
-function createWidget(
-  type: WidgetType,
-  overrides: Partial<Omit<UIWidgetBase, 'style' | 'type' | 'config'>> & { config?: Record<string, unknown>; style?: Partial<WidgetStyle> }
-): UIWidget {
-  const defaults = getWidgetDefaults(type);
-  const baseWidget: UIWidgetBase = {
-    id: nextId('widget'),
-    name: overrides.name || `${type}_${nextId('widget').slice(-4)}`,
-    type,
-    x: overrides.x ?? 50,
-    y: overrides.y ?? 50,
-    width: overrides.width ?? defaults.width,
-    height: overrides.height ?? defaults.height,
-    anchor: overrides.anchor ?? 'top_left',
-    style: { ...DEFAULT_WIDGET_STYLE, ...defaults.style, ...(overrides.style || {}) } as WidgetStyle,
-    visible: overrides.visible ?? true,
-    interactable: overrides.interactable ?? true,
-    parentWidgetId: overrides.parentWidgetId ?? null,
-    children: overrides.children ?? [],
-  };
-
-  const config = { ...defaults.config, ...overrides.config };
-
-  return { ...baseWidget, type, config } as unknown as UIWidget;
-}
